Only author can delete the question

+++ app/controllers/questions_controller.rb
@@ -1,5 +1,6 @@
 class QuestionsController < ApplicationController
   before_action :authenticate_user!, except: [:index, :show]
+
   def index
     @questions = Question.all
   end
@@ -14,9 +15,9 @@ class QuestionsController < ApplicationController
   def edit; end
 
   def create 
-    @question = Question.new(question_params)
+    @question = current_user.questions.new(question_params)
     if @question.save
-      redirect_to @question, notice: 'Your question successfully created'
+      redirect_to question_path(@question), notice: 'Your question successfully created'
     else
       render :new
     end
@@ -30,9 +31,13 @@ class QuestionsController < ApplicationController
     end
   end
 
-  def destroy 
-    question.destroy
-    redirect_to questions_path
+  def destroy
+    if current_user == question.author
+      question.destroy
+      redirect_to questions_path, notice: 'The question was deleted successfully'
+    else 
+      redirect_to question, alert: "You can't delete not your question"
+    end
   end
   
   private
